Audio: Lib AMC rework to avoid multiple AT write concurency
BZ: 6630

It was possible to send 2 AT cmd simultaneously and this is not
supported by the modem. To avoid any sync issue single write has been
implemented. Enhancment to recover if no response from the modem after 2
sec to a specific cmd, and retry to send it.
Add a dependency to patch#15922

Change-Id: I0314c03b0e1f9aac39b8e8c7f5a0d67d1edeb393
Signed-off-by: remi laudebat <remi.laudebat@intel.com>