[PORT FROM R2] [WORKAROUND][AUDIO] no Audio on CSV/VoIP call on BT with audience enabled
BZ: 17287

During a call, while trying to apply the preset 0x04 (dedicated to BT
acoustic), the audience chip could not output any sound.
For some reason (and until a new FW delivery from Audience), Audience chip
needs to have a 48khz clock on its Port A or C (in whatever I2S or PCM mode)
in order to apply the commands linked to the Preset 0x04.

Once again, this "ugly" patch is temporary until bug fixed by audience.
When a call BT is established, we have to switch either the modem (CSV) on again
in 48kHz mode or the MSIC PCM1 on again in 48kHz mode in order to apply the preset.
A sleep is kept before switching back the modem to 8kHz (or to switch off the MSIC PCM1 during VoIP)
in order to let audience running all its commands.
Swapping between VoIP / CSV is also taken into account, so the disable of the modem is done first.

Change-Id: I637fb236a49ad9f80b191627b3429c26ed578e53
Old-Change-Id: I576082f57e57393aad50b404e844e4d5b4bd7b20
Signed-off-by: Francois Gaffie <francoisx.gaffie@intel.com>
Reviewed-on: http://android.intel.com:8080/27799
Reviewed-by: Barthes, FabienX <fabienx.barthes@intel.com>
Tested-by: Barthes, FabienX <fabienx.barthes@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>