[PORT FROM R2] libamc: update for modem cold boot procedure
BZ: 17712

When the STMD wants to perform a modem cold boot, it will
broadcast the MODEM_COLD_RESET message to all clients.
At reception of this message, the client must acknowledge
the modem cold boot when it's ready by sending to the STMD
the MODEM_COLD_RESET_ACK message.

This patch adds the handling of this new STMD message and
the sending of the acknowledge.
Signed-off-by: Guillaume Lucas <guillaumex.lucas@intel.com>

Change-Id: Icbf6004006af7522b09cd84f282a55efc8b95999
Old-Change-Id: I5cea6c1dac087e3818d4b82a60c65901d0d71380
Reviewed-on: http://android.intel.com:8080/27857
Reviewed-by: Barthes, FabienX <fabienx.barthes@intel.com>
Tested-by: Barthes, FabienX <fabienx.barthes@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>