[PORT FROM R2][AUDIO] Audio Glitches at the establishment/release of voice call
BZ: 17698 16946

Glitch are listened at the establishement/release of call.
It was rootcaused that at initialisation of a call, the modem configures
its port by himself (AMC block) concurrently to PW processor.

It was etablished that we have to open the audio route only after the
modem has sent the xcallstat = Active (MT) or alerting (MO).

As a result, HAL had to listen to this xcallstat unsollicited AT command
to enable/disable audio routing for a call

The purpose of this patch is to deliver a new at-manager library
to provide services like sending unsollicited command, register callback
on stmd event, unsollicited command received.
It has also some adaptation changes within VPC, including a new set_mode
API to be able to set the current mode on VPC independantly of the
device.

Change-Id: I0539bfbbd76fd56d8bc07b1d17eb55924a9a3310
Orig-Change-Id: I3354fa73396dd6e09ef9059440466b1f8f3acff8
Signed-off-by: Francois Gaffie <francoisx.gaffie@intel.com>
Reviewed-on: http://android.intel.com:8080/30885
Reviewed-by: Barthes, FabienX <fabienx.barthes@intel.com>
Tested-by: Barthes, FabienX <fabienx.barthes@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>