[PORT FROM R2][AUDIO] Over consumption in CSV call
BZ: 19461

An overconsumption has been observed, about 30 mW, which is exactly
the impact of enabling I2S2 from modem side.

The purpose of this patch is to correctly managed the enable/disable
of the I2S2 port in case of mixing AND record (same issue)

This patch deals with AMC library (introduced a record off function)
and vpc.

Change-Id: Ib19bc30f2547c84878b2cd2f0e34aaef0fce1e52
Old-Change-Id: I106f3d059737fc7bd559f7e8595623ec970d9e6e
Signed-off-by: Francois Gaffie <francoisx.gaffie@intel.com>
Reviewed-on: http://android.intel.com:8080/31895
Reviewed-by: Barthes, FabienX <fabienx.barthes@intel.com>
Tested-by: Barthes, FabienX <fabienx.barthes@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>