[PORT FROM R2][AUDIO] No Audio on CSV call can occur
BZ: 19567

Audio can be lost on CSV call until next reboot.
It appears that the XCALLSTAT (recently changed to XPROGRESS) was never
received on Audio HAL while it was caught by RIL.
Logs showed that the registration to this AT command failed due to
concurrency and lack of protection within AMC library.

The purpose of this patch is to enforce security and protect members and
functions accessed by both worker thread and client thread.

Change-Id: I29773a1daa19fd25623250f5a3729faeb32d5df8
Old-Change-Id: I57a0b8cb1eb241974615d7e34a619e130c4e68d3
Signed-off-by: Francois Gaffie <francoisx.gaffie@intel.com>
Reviewed-on: http://android.intel.com:8080/32054
Reviewed-by: Barthes, FabienX <fabienx.barthes@intel.com>
Tested-by: Barthes, FabienX <fabienx.barthes@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>