[PORT FROM R2] [AUDIO] Register to XPROGRESS/XCALLSTAT may fail
BZ: 22246

In some case, register to XPROGRESS / XCALLSTAT may fail, as a result
audio during call will not be enabled.
In such case, RRIL is able to receive an ack to its registration to the same
message.
RRIL experts advised to use the same tty configuration as RRIL, as it seems to be
the only clue to solve this issue.
This patch also embeds cosmetic changes within traces to avoid confusion in such case
of registration NACK.

Change-Id: Ieed92a270eabd01a8f3ec8e37e211fbd2ce47733
Old-Change-Id: I563c91fc7d20b741d8aa29dd62aff0c2b068aa0f
Signed-off-by: Francois Gaffie <francoisx.gaffie@intel.com>
Reviewed-on: http://android.intel.com:8080/34459
Reviewed-by: Barthes, FabienX <fabienx.barthes@intel.com>
Tested-by: Barthes, FabienX <fabienx.barthes@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>