[AUDIO] Record not reactivated after an accessory switch
BZ: 21841

During an accessory change, the modem configuration is reseted and restablished
according to the new accessory. If a record was ongoing, the configuration is
overwritten and never restablished again. As a result, a starvation is observed
and the application closes the streams, even if the time counter is still increasing.

This patch checks upon the record flag within the vpc to restore the record path
after each CSV reconfiguration.

Change-Id: I9ab56d67bdf1d318a3807908ae3345e3c9e8fbb4
Signed-off-by: Francois Gaffie <francoisx.gaffie@intel.com>
Reviewed-on: http://android.intel.com:8080/34924
Reviewed-by: Tonnet, Pierre-louisX <pierre-louisx.tonnet@intel.com>
Reviewed-by: Centelles, Sylvain <sylvain.centelles@intel.com>
Tested-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>