hwc: fixed green bar flickering during playing AVC format clips
BZ: 23829

video driver updated the number of decoding buffer to 28 which
exceeds the max number of buffers in HWC. which may lead to green
bar shown on screen when a displaying buffer is unmapped when
HWC tries to shrink the used buffers when memory is low.

This patch increased the number of buffers to 30 in HWC, so that
the AVC decoding won't lead to any buffer unmap operation during
video playback.

This patch also redefined the buffer number for sprite planes and
overlay planes, and stabilize the buffer invalidate operation.

Change-Id: I70c6242d19593707d524eb74f4a353416c97f400
Signed-off-by: Jackie Li <yaodong.li@intel.com>
Reviewed-on: http://android.intel.com:8080/35539
Reviewed-by: Ding, Haitao <haitao.ding@intel.com>
Tested-by: Ding, Haitao <haitao.ding@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>