[PORT FROM R2][AUDIO] Audio Lost when calling conf bridge after a restart of AudioHAL
BZ: 22586

If a crash occurs on AudioHAL, AudioHAL is restarted. Registration
to XCALLSTAT and XPROGRESS needs to be done again, and it trigs twice
the eventhread. As a result, processing of the tosend queue of AT command
is done twice, breaking the serialization of the communication with the
modem. Only one ACK is received, loosing the registration of XCALLSTAT.
For use cases in which modem does not provide XPROGRESS info, the
audio is lost during call (as for example calling conference bridge)

This patch ensures of the serialization of the comm with the modem on
the tty.

Change-Id: If1e408c909068db0dd696328a937b096daa24ec6
Orig-Change-Id: Idf025cace26dd04cb597b817546cbc8bbf1b0aa0
Signed-off-by: Francois Gaffie <francoisx.gaffie@intel.com>
Reviewed-on: http://android.intel.com:8080/35753
Reviewed-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Tested-by: Barthes, FabienX <fabienx.barthes@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>