[AUDIO][GI] No Audio on CSV Call
BZ: 24601

GI platform does not embed Audience Chip.
The SW from R3 was used without any changes. However, when initializing
VPC, audience cannot be initialized, and init is failing.
As a result, this HW module is closed from Audio HAL.

The purpose of this patch is to remove the Audience CFLAGS from VPC
android makefile.

Change-Id: I6e84224f373db88cac00d4a7159759599f23bf55
Signed-off-by: Francois Gaffie <francoisx.gaffie@intel.com>
Signed-off-by: Louis Le Gall <louis.le.gall@intel.com>
Reviewed-on: http://android.intel.com:8080/36512
Reviewed-by: Seibel, Eric <eric.seibel@intel.com>
Tested-by: Barthes, FabienX <fabienx.barthes@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>