hwc: change sprite format from BGRA8888 to BGRX8888 for benchmark test
BZ: 25660

this patch provides a work around in hwc for glbenchmarkes1.0 test. As the BGRA8888 setting
to Display controller should be pre-multiplied, but the hardware seems to do the multiply of
alpha channel. Now this patch just set it to be BGRX8888 to ignore alpha channel. Need to
confirm with H.W. designer about the effect of BGRA8888 reg setting.

Change-Id: Idde30f45ab0e3db9d8602015e08728c78f2bc99e
Signed-off-by: Lei Zhang <lei.zhang@intel.com>
Reviewed-on: http://android.intel.com:8080/36951
Reviewed-by: Xu, Randy <randy.xu@intel.com>
Tested-by: Tong, BoX <box.tong@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>