hwc: signal render status in surface payload
BZ: 25331

Use the new payload flag to signal whether a surface has been
completely render or not.
Current implementation is used for Widi only, but same approach can
be taken for HDMI or HW overlay
The payload buffers attached to the gralloc buffers are mapped at when
they are collected before the transition to extVideo Mode and  unmaped
when we go back to clone mode.

Change-Id: I4a0bad46a20a908e43a92d1568bfe4e2fd82788c
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Reviewed-on: http://android.intel.com:8080/37283
Reviewed-by: Alanen, Juha <juha.alanen@intel.com>
Reviewed-by: Xu, Randy <randy.xu@intel.com>
Tested-by: Sun, Hang L <hang.l.sun@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>