[PORT FROM R2][AUDIO] Audio HAL ATManager to request modem reset on error
BZ: 18904

In case of error (like pollerr, hangup, ...), ATManager instance within
the Audio HAL will just close its tty. Until a modem Down/up event is triggered,
audio HAL will not be able to receive XCALLSTAT/XPROGRESS event
from the modem.

This patch adds the capability to perform a clean up request to stmd
(modem reset) when an error is detected after expiration of a
timeout and if the modem is still up.

Change-Id: If4b4618bd2e3a5060766fec2625fefead145c4aa
Orig-Change-Id: I739ba9864c73dddc3fec45c8525c981dad7c4215
Signed-off-by: Francois Gaffie <francoisx.gaffie@intel.com>
Reviewed-on: http://android.intel.com:8080/42422
Reviewed-by: Barthes, FabienX <fabienx.barthes@intel.com>
Tested-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>