widi: Removed the orientation changed when in extended mode.
BZ: 32050

The orientation of the video was being changed when in exteded mode, which is unnecessary in exteded mode.
The impact was that with 3rd party app storm player, the video on remote display was in portrait mode with
screen divided into two. This patch removes the orietation changed when switching to exteded mode.

Change-Id: I83cbe8c4c8b6b3216a2330c7a610c70fb8c557e4
Signed-off-by: pragnesh ramolia <pragneshx.r.ramolia@intel.com>
Reviewed-on: http://android.intel.com:8080/43619
Reviewed-by: Tao, Tao Q <tao.q.tao@intel.com>
Reviewed-by: Hu, Jason <jason.hu@intel.com>
Reviewed-by: Zhu, Tianyang <tianyang.zhu@intel.com>
Reviewed-by: Xu, Randy <randy.xu@intel.com>
Tested-by: Sun, Hang L <hang.l.sun@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>