[AUDIO] Lost of Audio on CSV CALL while BPLogs activated
BZ: 32082

DLCI open is asynchronous. When BP logs are activated, DLCI
open can take until 50ms to be achieved. If a write is performed
before DLCI ACK sent by the modem, an error is now returned to the
client (ATManager). As a result, a write can now fails. If it fails,
ATManager needs to retry to send the command.
It will try 5 times every second. If it fails after this procedure,
a request cleanup will be performed.

Change-Id: Ibfb72842ec352629bdd6425b78a299fb71ede5da
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>
Reviewed-on: http://android.intel.com:8080/43755
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Tested-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>