Do not send khanle with value 0 to widi
BZ: 32270

When there's a phone call comming, decoder is destroyed and psb-video
clears payload buffer which contains field khandle. The fix checks
whether the value of khandle is 0, if it is, switch to clone mode and
do not send the khandle to widi.

The patch also adds lock to protect extended payload map and etc.

Some compiling warning are also fixed.

Change-Id: I7e70d5a18917940c72bd38358390c08179539c1d
Signed-off-by: ttao1 <tao.q.tao@intel.com>
Reviewed-on: http://android.intel.com:8080/44321
Reviewed-by: Rogers, BrianX E <brianx.e.rogers@intel.com>
Reviewed-by: Balguri, Mamatha <mamatha.balguri@intel.com>
Reviewed-by: Veeramani, Karthik <karthik.veeramani@intel.com>
Reviewed-by: Xu, Randy <randy.xu@intel.com>
Tested-by: Sun, Hang L <hang.l.sun@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>