[AUDIO] TTY does not work correctly
BZ: 31131

TTY does not work correctly. Advanced debug showed that audience
is not always in pass trough when TTY mode is on.
It is due to race condition as setParameter API of Audio HAL and
setStreamParameter can be called in any order to respectively
configure the tty mode and change the output device on jack insertion.

Audio HAL cannot rely on the right order. Need to reconsider the routing
in case of setting new tty mode to take it into account.

This patch also corrects a bug within the check of tty_call (considered
as a boolean whereas it can takes several values)

Change-Id: Iee7ba6909ec1296fef9c4fbd47fbd49a39825c0c
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>
Reviewed-on: http://android.intel.com:8080/46403
Reviewed-by: Becker, VincentX <vincentx.becker@intel.com>
Reviewed-by: Schmuck, GuillaumeX <guillaumex.schmuck@intel.com>
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Tested-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>