Audio-HAL: Music cannot be heard from HDMI after dismiss the alarm.
BZ: 26414

During HDMI open, the delay in processing is caused by the Silence buffer
feature in driver code which is provided to overcome hardware limitation.
The processing of the open request for HDMI is done at the interrupt boundary
in driver code, which takes max of 23ms. So any open request would be responded
by a -EAGAIN till this interrupt boundary. The ALSA layer returns -EBUSY when
driver returns -EAGAIN.
This error is not handled in "s_open()" and the requests are sent to open the
device after truncating "devName" at "_". As these unknown device names are not
recognized by driver, the generic device (IHF) is opened.
With this fix, -EBUSY is handled by sending repeated requests for MAX 6 times,
without truncating the "devName", after a delay of 10ms each time.

Change-Id: Ieb66d02db7efa9510aeb3a91c526b0452c2380c2
Signed-off-by: gbhakteX <gurudattax.bhakte@intel.com>
Reviewed-on: http://android.intel.com:8080/46412
Reviewed-by: Sikkandar D, Madar <madar.sikkandar.d@intel.com>
Reviewed-by: Kandasamy, Muthukumar <muthukumar.kandasamy@intel.com>
Reviewed-by: Sameullah, MazharX <mazharx.sameullah@intel.com>
Reviewed-by: Gaffie, Francois <francois.gaffie@intel.com>
Reviewed-by: Centelles, Sylvain <sylvain.centelles@intel.com>
Reviewed-by: Gupta, ArvindX K <arvindx.k.gupta@intel.com>
Reviewed-by: Hibare, PramodX <pramodx.hibare@intel.com>
Tested-by: Hibare, PramodX <pramodx.hibare@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>