[AUDIO][LEX] VoIP path moved to media route ie PCM2 MSIC
BZ: 31748

As Audience chip is not available on Lex, and as LPE needs to
be used to perform some processing instead, VoIP path must
be moved to Media Route, ie from IA I2S2 to MSIC PCM2. (This change
only affects when playing on Builtin accessories, not on BTSCO)

This patch force the vpc to ignore INCOMMUNICATION mode. BT
path will remains available in NORMAL mode
when VoIP is playing on BTSCO.

Change-Id: I5a036963b333d41e5914be1408feda0fbd457f3e
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>
Reviewed-on: http://android.intel.com:8080/46967
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Tested-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>