Enabling RELRO: make all hw_module_t instances non-const
BZ: 35225

Hardware modules which are eventually loaded by libhardware have
a 'descriptor' in the form of a hw_module_t data structure. This
structure is generally declared as 'const' in the various hardware
lib modules inside Android. While this is conceptually OK, there is
a problem with the 'load' function in libhardware (hardware.c)
which tries to write the 'dso' field of the hw_module_t structure
after loading a library (ignoring its 'const' attribute). After
enabling RELRO, hw_module_t is part of the RELRO segment (thus
declared as read-only) and any program that uses 'load' from
libhardware terminates with a segmentation fault error. This patch
removes the 'const' part of hw_module_t instances, which moves
hw_module_t out of the RELRO area and enables correct operation.

Change-Id: I815e1b78c9c49996d8b28d0874fab3cdf3b3af70
Signed-off-by: lwan89x <liangx.wang@intel.com>
Reviewed-on: http://android.intel.com:8080/47907
Reviewed-by: Qin, Xiaokang <xiaokang.qin@intel.com>
Reviewed-by: Zhang, Hongyu <hongyu.zhang@intel.com>
Tested-by: Li, XiaojingX <xiaojingx.li@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>