During simultaneous Playback and Capture use-case causing broken pipe errors (-EPIPE)
BZ: 33738

The current period size is 46ms in HAL. FW also reads a
full period at a time. Because of this when the first HAL
write happens, FW reads the full period and thus alsa
have a empty ring which resulted in the EPIPE error.
Start threshold is now changed to full buffer instead
of a single period for media playback so that first FW
read happens only after 2 period HAL writes.

Change-Id: I80ec119aede2ff318b99aad76bf1f39a33d8552e
Signed-off-by: Muthukumar K <muthukumar.kandasamy@intel.com>
Reviewed-on: http://android.intel.com:8080/47937
Reviewed-by: Sikkandar D, Madar <madar.sikkandar.d@intel.com>
Reviewed-by: Sameullah, MazharX <mazharx.sameullah@intel.com>
Reviewed-by: Centelles, Sylvain <sylvain.centelles@intel.com>
Reviewed-by: Gupta, ArvindX K <arvindx.k.gupta@intel.com>
Reviewed-by: Krishna, Samaga <samaga.krishna@intel.com>
Reviewed-by: Hibare, PramodX <pramodx.hibare@intel.com>
Tested-by: Hibare, PramodX <pramodx.hibare@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>