[Audio] Change DEVICE_OUT from earpiece to speaker.
BZ: 39477

In vpc_route, in case of using BT_SCO, change the pcm_enable
used to open a MSIC device on PCM1 from DEVICE_OUT_EARPIECE
to DEVICE_OUT_SPEAKER as there is no Earpiece on tablet.
Aim is to enable a clock, not to play audio on it.
This will not affect other platform as SPEAKER is on all platform.

Change-Id: Ia6528f63066fffe5ccdd38781ad40f737d53f13d
Signed-off-by: Renaud Viry <renaudx.viry@intel.com>
Reviewed-on: http://android.intel.com:8080/50918
Reviewed-by: Bellouard, Jean-marie <jean-marie.bellouard@intel.com>
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Tested-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>