[WORKAROUND][TO BE REVERTED] No Audio on VoIP with BT without acoustic
BZ: 33934

When using a BT without acoustic, audio path cannot be established
during VoIP call.

As on CTP, the cirus chip Voice Port is slave, the clock of the cirus
cannot be switched on temporary to set the audience profile.

As know, audience has a known issue when switching from 48k to 8k mode.
It has to have a PCM clock at 48k still active to apply its 8k profile.

This bug has been reported to audience.

This patch implements an ugly WA: use the modem clock to switch to
BT audience profile.

Note that if a tablet is done on CTP base, without modem, we will face
this issue again....without workaround possible this time..

Change-Id: Ibf614194fcf26142b2ef57769161d33c9f9a7361
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>
Reviewed-on: http://android.intel.com:8080/50871
Reviewed-by: Becker, VincentX <vincentx.becker@intel.com>
Reviewed-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Tested-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>