[PORT FROM R3][Audio] AuViD not working properly
BZ: 36199

AuViD is the PC based tool provided by Audience to communicate with
the es305b chip for tuning purpose. The communication is done using
the ad_proxy which forward AuViD commands to the es305b chip using
i2c bus.

The tool cannot be used because of incoherent values reported. This
issue is root caused by the acoustic module of the VPC, which never
read response from the es305b chip when it has sent a command to it.

Because of that, the es305b i2c tx fifo is full of old commands
responses not flushed, and these old commands response are read by
AuViD tool instead of the responses it is supposed to read.

The patch update the acoustic module to make it compliant with the
es305b protocol: for each command sent, the ack is readed. That will
avoid to corrupt the es305b tx fifo with old unpoped response.

A second change is needed to make the acoustic module compliant with
the proxy: it must not suspend the Audience chip since the proxy may
need to access to the Audience chip at any time.

A property is added to configure the acoustic in a such mode: when
the persisten property persist.audiocomms.vp.tuning.on is set to true
the acoustic module never suspend Audience.

Change-Id: I4dffc4e1bb7811d841be943260f4986b49684bcf
Orig-Change-Id: I3f9ff607d4ccd6eeb928cfb9f621f9572272dd5e
Signed-off-by: Mathieu Soulard <mathieu.soulard@intel.com>
Reviewed-on: http://android.intel.com:8080/54712
Reviewed-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Tested-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>