[PORT FROM R3] [AUDIO] Crash in AT Manager
BZ: 40621

Even if the issue was raised with ENZO HW, the guilty code is
common and required to be patched.
When the modem is too long to answer (> client timeout), the
client semaphore is posted, and the client will return and send
quickly the following AT Command. The AT manager did not cancel
this request, leaving an orphan ATCmd pointer, that can lead to
a crash, even a seg fault.

The AT Manager has a protection mechanism: after 10 sec, if no
answer is received from the modem, it will reset the modem.
The timeout protection from the client is useless as the client
cannot recover from this error.

Note that for commodity reason, the recover mechanism of the
ATManager can be disabled using Android Property.

Change-Id: I60baaf15b79e2befe0f9fd26678a3f822512d2c3
Orig-Change-Id: I3511977689a051f9361034ccea73acc0b44534cc
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>
Reviewed-on: http://android.intel.com:8080/58322
Reviewed-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Tested-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>