[PORT FROM R3] Audio:  [Audio_Comms] Fix for audio media playback latency in normal mode
BZ: 42900

The issue is introduced by patch 51758 and 51757 where the latency is increased due to buffer size increase.
The patch consists in reverting to 2 periods of 46ms in normal mode, and changing the buffer size
at flinger interface rather than HAL, thus not impacting latency.
Tested OK with initial issue 31926, and CTS test described in 42900

Change-Id: I910d1e060c3e8dea1d5cf5cfd892b0698b48a062
Orig-Change-Id: Id1be2b46b1251af01ecb6a5516e718a74bbe3437
Signed-off-by: Guillaume Schmuck <guillaumex.schmuck@intel.com>
Reviewed-on: http://android.intel.com:8080/58323
Reviewed-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Tested-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>