[PORT FROM R3] [Audio] Do not compile Audience Proxy on LEX
BZ: 43510

LEX does not include Audience chip. All code related to
Audience is not compiled on LEX, except the Audience proxy
which has not been removed.

Because of that, Audience proxy is built for LEX and copied
on LEX debug images.

This patch avoids the compilation of the Audience proxy for LEX.

Signed-off-by: Mathieu Soulard <mathieu.soulard@intel.com>

Change-Id: Ieea49706774b045a40683b14b302e401a556e248
Orig-Change-Id: I905ff24a983c517d8b1d0742ccbcf3729e3c7679
Reviewed-on: http://android.intel.com:8080/56967
Reviewed-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Tested-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>