[PORT FROM R3] [AUDIO] Lost of 2 seconds of call when accepting it
BZ: 29679

The setMode in call was recently changed and is sent earlier.
The sequence is:
-setMode In call from the policy
-ModemCallStatus=true received from AT Manager
-setStreamParameter with the device to use received from
the policy.

As a result, when receiving the call status, the route is established
on the wrong device. When the setStreamParameter is received, a re-
routing is performed, increasing by around 800ms the establishment
of the audio call.

This patch ensures unecessary steps are not performed at VPC level.

Change-Id: I1922dffe440888d338ed671589b9ae220bf7f046
Orig-Change-Id: I1adaf82a618323799ec2d86d276294b56886d3f4
Old-Change-Id: I6effb8bb4498065d9c8492a8248e73468742e961
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>
Reviewed-on: http://android.intel.com:8080/60219
Reviewed-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Tested-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>