[Audio] NB/WB support
BZ: 42137

In order to support narrow & wide bands specific Audience tuning, the
VPC needs to know which band is currently in use for CSV & VOIP calls.

This pathc introduces a new interface to the VPC allowing to let it know
which band is used for CSV and for VOIP.

Depending on that information providen by upper layers, the VPC will
configure Audience for narrow or wide band.

Change-Id: I05573994e6c67ef52f30530db6ec24dbc1fdafcc
Signed-off-by: Mathieu Soulard <mathieu.soulard@intel.com>
Reviewed-on: http://android.intel.com:8080/57737
Reviewed-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Tested-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>