[Audio] Audio WB/NB Handover management
BZ: 42650

This patch introduces Wide Band and Narrow Band support.

The AudioATManager now subscribes to the Modem notifications
about PCM change to be notified of codec switch.

This information is retreived by unsollicited AT command XDRVI
and forwarded to the observer clients.

This patch updates the ATCommand and UnsollicitedATCommand design
introducing the handling of ATCommand answer and UnsollicitedATCommand
notification from the Modem.

Change-Id: Ibb65a4bfe4132e65398dea6a9ac4c17cc3fd9697
Signed-off-by: Mathieu Soulard <mathieu.soulard@intel.com>
Signed-off-by: Vincent Becker <vincentx.becker@intel.com>
Reviewed-on: http://android.intel.com:8080/58373
Reviewed-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Tested-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>