[VIBRA] Adapt Intel vibrator HAL into a hardware module
BZ: 49760

Intel vendor implementation of vibrator needs to be done inside a
hardware module, so that it can overload the default AOSP implementation.
This is needed as the Intel vibra uses the parameter framework to get
the path of the vibrator system file. The module name becomes vibrator.<target_device>.so.

Change-Id: I88a182c3e71da472b72c2230e06d01a84c671154
JB-integ-Change-Id: Ib4482e763d81f375e76e543652db13e488ea8ea0
Signed-off-by: Vincent Becker <vincentx.becker@intel.com>
Reviewed-on: http://android.intel.com:8080/62343
Reviewed-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Tested-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>