[PORT FROM R3] WiDi: Video screen is messy after ending a call over WiDi
BZ: 50653

disablePlane call in stopstreaming is causing the WiDiPlane state to switch to INITIALIZED from ACTIVE.
Which causes the HWC to switch the interface and when call ends this causes side effects as buffers are
not matched.

Introduced a parameter in disablePlane to check  whether widi is still connected but not streaming.
Avoiding flip in the case where we are connected and not streaming.

Change-Id: I6180a434b2634adfd721fa43b0f752095f210877
Signed-off-by: Aslam Padath Peedikayil Abdul Rahim <aslamX.padath.peedikayil.abdul.rahim@intel.com>
Reviewed-on: http://android.intel.com:8080/64711
Reviewed-by: Padath Peedikayil Abdul Rahim, AslamX <aslamx.padath.peedikayil.abdul.rahim@intel.com>
Reviewed-by: Tao, Tao Q <tao.q.tao@intel.com>
Reviewed-by: Xu, Randy <randy.xu@intel.com>
Tested-by: Post, DavidX J <davidx.j.post@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>