[PORT FROM R3] hwc: fix video rotation flicker on LEX
BZ: 52861

By reverting previous overlay flip and post2 sequence patch, it shows
video flicker issue on LEX. This is the patch to fix the issue on Lex.
Meanwhile it should not cause other regression on ENZO.

The rootcause of the flicker issue is because the overlay flip flag is
wrongly set to be WMS_NEEDED in rotation cases, which cause write_mem_start
happens twice in driver.

Change-Id: Iaeb75a06fbfeef49f10b79028eac9b3426913c90
Signed-off-by: Lei Zhang <lei.zhang@intel.com>
Signed-off-by: Tong, Bo <box.tong@intel.com>
Reviewed-on: http://android.intel.com:8080/63075
Reviewed-by: Xu, Randy <randy.xu@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>