hwc: fix fountain apk display black issue
BZ: 49014

This is the patch fixes the fountain apk display black on Jellybean.
It is due to the setzOder function is commented, which causes the
framebuffer format setting is wrongly set to be RGBA. And then it will
do blending according to alpha channel. With no overlay layer, no
need to do so and the format should be RGBX.

Change-Id: I48fa9db91fb4681ec0120e8e15d98c70b3ab38b9
Signed-off-by: Lei Zhang <lei.zhang@intel.com>
Reviewed-on: http://android.intel.com:8080/63716
Reviewed-by: Hu, Jason <jason.hu@intel.com>
Tested-by: Shi, PingX <pingx.shi@intel.com>
Reviewed-by: Xu, Randy <randy.xu@intel.com>
Tested-by: Xu, Randy <randy.xu@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>