[AUDIO] New audio media buffer size and latency values
BZ: 50161

During VoIP call, some cuts (blanks) can be heard in UL.
Analisys with pyTimeChart showed a scheduling issue between
SIP Network thread, Audio In and Audio Out. Audio out wrote
a buffer to the HAL every 80ms. Reducing to 23 ms between
AudioFlinger in Audio HAL, and 4*20ms periods in ALSA in input
and output allows to have a fluent scheduling and not miss
the train when transmitting a new buffer to the distant.

Change-Id: I2a984ad5b5ff7fd7906cebc009d6706c8b29b0e9
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>
Reviewed-on: http://android.intel.com:8080/64869
Reviewed-by: Di Folco, NeilX <neilx.di.folco@intel.com>
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Reviewed-by: Quervel, Thierry <thierry.quervel@intel.com>
Tested-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>