widi: Handle display on/off while widi is switching between extended/clone mode
BZ: 54635 56526

Currently, there are several issues with turning on/off mipi in case of widi.
1. The Mipi is not turned off when the player is started without any other layer (i.e. controls) on top of video
2. The mipi is not turned back on if there is a phone call while in ext video mode
3. The mipi is not turned back on sometimes if home button is pressed while in ext video mode
4. The mipi is not turned back on if widi streaming is stopped while in ext video mode

All these issues are caused by combination of widi state flags and some other HWC flags are not
in the right state at the right time to turn on/off the mipi.

This patch changes the widi status everytime widi switches between clone and ext video mode. Also the mipi
will be turned off/on as necessary based on the widi status change and widi state information.

Signed-off-by: Mamatha Balguri <mamatha.balguri@intel.com>
Change-Id: I2ffc02fd8fb01da45b1427fc348044ac79e131f2
Reviewed-on: http://android.intel.com:8080/65711
Reviewed-by: Padath Peedikayil Abdul Rahim, AslamX <aslamx.padath.peedikayil.abdul.rahim@intel.com>
Reviewed-by: Tao, Tao Q <tao.q.tao@intel.com>
Reviewed-by: Zhang, Lei <lei.zhang@intel.com>
Reviewed-by: Zhu, Tianyang <tianyang.zhu@intel.com>
Reviewed-by: Xu, Randy <randy.xu@intel.com>
Tested-by: Post, DavidX J <davidx.j.post@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>