[Audio] Audience band-specific preset is not applied during voice call
BZ: 56499

During voice call, the modem signals the codec changes which allow
to select WB or NB tuning acording to the current codec.

The NB or WB specific Audience preset should be selected but it fails
due to a wrong condition.

This patch updates the wrong condition to allow the preset selection.

Change-Id: I9db8295fbb3634e329cbdef588b29921235f97f1
Signed-off-by: Mathieu Soulard <mathieu.soulard@intel.com>
Reviewed-on: http://android.intel.com:8080/65619
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Tested-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>