[Audio] TTY enabled but not used disables acoustic processing
BZ: 56883

The TTY devices requires acoustics to be disabled. But acoustic
processing is still required on other accessories like headphone,
earpiece or speaker even if TTY is enabled.

This patch disable acoustic only if the TTY is enabled AND in-use,
by checking that the current accessory is wired headset, since the
TTY devices is detected as a wired headset.

Change-Id: I93cb58e2a8d5585ecf11b857c8140a6ba6164288
Signed-off-by: Mathieu Soulard <mathieu.soulard@intel.com>
Reviewed-on: http://android.intel.com:8080/65966
Reviewed-by: Gaffie, Francois <francois.gaffie@intel.com>
Reviewed-by: Bellouard, Jean-marie <jean-marie.bellouard@intel.com>
Reviewed-by: Centelles, Sylvain <sylvain.centelles@intel.com>
Tested-by: Barthes, FabienX <fabienx.barthes@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>