[Audio] camera shutter sound and voice call in call+SCO
BZ: 59242

When taking a picture while in a call the audio stream is
heard over the speakerphone for a second or two
when the shutter sound is played.

The solution is to set the device as BT SCO+ Speaker
in the policy.
ALSA will then recognize that it must be played as normal mode.
The Voice call will be unrouted and the shutter sound is played
in Speaker.
Then here we should not set the sample rate to 8 kHz.

This patch is a cherry-pick from patch 66796 from R3-stable.

Change-Id: I49a571a34da8994dbbfcf5a7b81f9cf1a92573ab
Orig-Change-Id: I248310062fdfe69b676176d9f65670856980d65a
Signed-off-by: pafonsoX <philippex.afonso@intel.com>
Reviewed-on: http://android.intel.com:8080/68956
Reviewed-by: Becker, VincentX <vincentx.becker@intel.com>
Reviewed-by: Gaffie, Francois <francois.gaffie@intel.com>
Reviewed-by: Di Folco, NeilX <neilx.di.folco@intel.com>
Reviewed-by: Centelles, Sylvain <sylvain.centelles@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>