Add property to change control modem AT channel
BZ: 51212

Replace the macro AUDIO_AT_CHANNEL_NAME contain the device file name
by a macro AUDIO_AT_CONTROL_CHANNEL_PROPERTY_NAME contain the property name of this device file name.

In vpc use this new macro instead of the old one, get the value of the property and use it for the device file name.

To implement a modem simulator for HAL test project, we need to change dynamically the device file name of the primary and control TTY channel.
For the primary channel (/dev/gsmtty20) the job is already done.

Change-Id: I3dc8f0de169598b64b8a4fecf7b62763396b15f4
Signed-off-by: Jeremie Graulle <jeremiex.graulle@intel.com>
Reviewed-on: http://android.intel.com:8080/65462
Reviewed-by: Soulard, Mathieu <mathieu.soulard@intel.com>
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Tested-by: Barthes, FabienX <fabienx.barthes@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>