[WORKAROUND] [AUDIO] Set playback period time according to the sample rate
BZ: 56773

For Widi output device and 48Khz sample rate, take half of the period
time used for other output devices as buffering needs to be faster to
avoid Widi video stuttering.
Period time is calculated now from the sample rate of the platform and
rounded up to a 16-frames boundary. Hence, the audio buffer is always
written at the right pace, limiting possible underruns or overflows.

Change-Id: Idbdb7e7871197bd32914b6ea82d59930ec380403
Signed-off-by: Vincent Becker <vincentx.becker@intel.com>
Reviewed-on: http://android.intel.com:8080/67744
Reviewed-by: Bellouard, Jean-marie <jean-marie.bellouard@intel.com>
Reviewed-by: Centelles, Sylvain <sylvain.centelles@intel.com>
Reviewed-by: Gaffie, Francois <francois.gaffie@intel.com>
Reviewed-by: Quintero, Jorge <jorge.quintero@intel.com>
Reviewed-by: Sikkandar D, Madar <madar.sikkandar.d@intel.com>
Tested-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>