HWC: Fix one incorrect YUV and RGB layers blending issue
BZ: 61388

This bug is a case that YUV layer is in the middle of the layer list
(layer list like RGB/YUV/RGB). YUV layer covers one RGB layer under
it. HWC needs blending overlay and sprite plane; overlay has to be set
to the bottom plane; so it changes the origal Z order of layer list,
and causes some messy show on the YUV region.
For such case, fall back to SurfaceFlinger to do composition.

Change-Id: If4c35050f2dfc8a4dfec5004eba2da0eb78ce891
Signed-off-by: Lei Zhang <lei.zhang@intel.com>
Reviewed-on: http://android.intel.com:8080/70984
Reviewed-by: Mao, Marc <marc.mao@intel.com>
Tested-by: Tong, BoX <box.tong@intel.com>
Reviewed-by: Tong, BoX <box.tong@intel.com>
Reviewed-by: Xu, Randy <randy.xu@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>