HWC: Fix a flash of black on the screen during WV playback
BZ: 53771

This is the patch to fix flash of black during WV playback while holding
down the home button to switch apps. One cause of flash is due to overlay
disabled in the process, so add protected layer check in isScreenshotActive.
The other cause is that there are layer list cases of video layer is
between two RGB layer, and blending of two RGB layer in SF may cover the
region of overlay.

This patch re-enables the coverage check and HWC_HINT_CLEAR_FB.

Change-Id: I8f0d28024aa04070e1151fcf017955839c38983c
Signed-off-by: Lei Zhang <lei.zhang@intel.com>
Reviewed-on: http://android.intel.com:8080/71414
Reviewed-by: Saffores, Ryan D <ryan.d.saffores@intel.com>
Tested-by: Shi, PingX <pingx.shi@intel.com>
Reviewed-by: Xu, Randy <randy.xu@intel.com>
Tested-by: Xu, Randy <randy.xu@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>