HWC: fix NULL bufferHandler when overlay context flip is bypassed
BZ: 64042

The crash is due to the NULL pointer of bufferHandler passed to Post2.
This is the patch to fix NULL bufferHandler on the case of overlay context
flip is bypassed. When flag set to be DELAY_DISABLE or widi active, the
overlay context flip now will be bypassed, and overlay will be disabled.
Under such case, it should not set the bufferHandler of the plane any more.

Change-Id: I45b80eaef850a62b6c8e3f52f95e6f7b5a6315ff
Signed-off-by: Lei Zhang <lei.zhang@intel.com>
Reviewed-on: http://android.intel.com:8080/71627
Reviewed-by: Xu, Randy <randy.xu@intel.com>
Reviewed-by: Mao, Marc <marc.mao@intel.com>
Tested-by: Tong, BoX <box.tong@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>