[AUDIO] Capture uplink after modem voice processings
BZ: 65122

On MFLD, +6dB is applied by the modem on UL in order to reach 3GPP specs.
But the Voice call record is not using the right probing point : this +6dB
and other modem voice processings are not included in the recorded uplink.
So all voice call record must be probed after algo processing in modem
so we will have the closest uplink as what is sent to the network.
Thus this requirement does not depend on Audience.
But there is an issue with the modem 6360 (clovertrail 1.5) so this probing
point will not be use by this modem : the bug 66682 has been created to
solve this issue.

Change-Id: I52c0a670d04367a2816ed422ab0d29e49ecf52f6
Signed-off-by: pafonsoX <philippex.afonso@intel.com>
Reviewed-on: http://android.intel.com:8080/72754
Reviewed-by: lab_cactus <lab_cactus@intel.com>
Reviewed-by: Di Folco, NeilX <neilx.di.folco@intel.com>
Reviewed-by: Soulard, Mathieu <mathieu.soulard@intel.com>
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>