HWC: fix HDMI does not work after first time flash
BZ: 69058

This is the patch to fix HDMI does not work after first time flash.
It is due to the MDS service is not found during HWC intilization.
MDS service now is adjusted to be loaded before SurfaceFlinger, the
issue maybe due to services are loaded in parallel. Fix it by trying
more times to get MDS service.

Change-Id: I6cced70e0a05088f99e6ec4d0cab230f4d08f463
Signed-off-by: Zhang,Lei <lei.zhang@intel.com>
Reviewed-on: http://android.intel.com:8080/76872
Tested-by: Tong, BoX <box.tong@intel.com>
Reviewed-by: Tong, BoX <box.tong@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>