[AUDIO] Support new audio devices enum values introduced in JB MR1 code drop
BZ: 71815

Today, each audio device is identified by a bit in a 32 bit field with
16 bits for output devices and 16 bits for input devices.
So the number of output devices is now exhausted i.e. we can't add new output devices.

So in order to add new devices, a new convention for audio device
identification is defined, allowing 31 different input and output devices.
Bit 31 of the bit field is reserved and when set indicates an input device.

Change-Id: I2d45471bf349ea925f0958157e961548385c7d0f
Signed-off-by: ndifolcx <neilx.di.folco@intel.com>
Reviewed-on: http://android.intel.com:8080/77835
Reviewed-by: Barthes, FabienX <fabienx.barthes@intel.com>
Reviewed-by: Gaffie, Francois <francois.gaffie@intel.com>
Tested-by: Barthes, FabienX <fabienx.barthes@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>